# openapi.yaml
openapi: 3.0.3
info:
  title: 31for31 JustWatch Scraper API
  description: >
    REST API in front of the Supabase film catalogue (the source of truth).
    The `/films` endpoints read and write films and their `justwatch_url`
    overrides, and refresh each film's current streaming offers (free /
    subscription / rent / buy) by scraping JustWatch into Supabase. The
    `/api/scrape` endpoints are a stateless, ad-hoc scrape that persists
    nothing -- handy as a dry run. See the [README](README.md) for how the
    scraping itself works.
  version: "1.1.0"
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://31for31scraper.vercel.app
    description: Production (Vercel)
  - url: http://127.0.0.1:5000
    description: Local development
...
paths:
...
  /api/scrape:
    get:
      operationId: scrape
      tags: [Scrape]
      summary: Scrape a single film's current JustWatch streaming offers
      description: >
        Scrapes one film ad hoc and returns a `FilmResult`. For more than
        one film in a single request, use `POST /api/scrape` instead --
        it's the same underlying scrape, just for a caller-supplied list,
        with far fewer round trips.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: title
          in: query
          required: true
          description: The film to scrape.
          schema:
            type: string
          example: "The Strangers (2008)"
        - name: url
          in: query
          required: false
          description: >
            Scrapes this exact JustWatch URL directly instead of guessing
            the slug from the title -- this is how you'd try out a
            `justwatch_url` override before using it elsewhere. Must be an
            `https://www.justwatch.com/...` URL -- anything else is
            rejected (this endpoint won't fetch arbitrary URLs on your
            behalf).
          schema:
            type: string
            format: uri
          example: "https://www.justwatch.com/us/movie/event-horizon-1997"
      responses:
        "200":
          description: Scrape succeeded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FilmResult"
              example:
                title: "The Thing From Another World"
                url: "https://www.justwatch.com/us/movie/the-thing-from-another-world"
                service:
                  - name: "The Roku Channel"
                    type: "free"
                    price: null
                    currency: "USD"
                    link: "https://therokuchannel.roku.com/details/..."
                    icon: "https://images.justwatch.com/icon/76972041/s100/rokuchannel.webp"
        "400":
          description: The `title` query param was omitted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: '"title" query param is required'
        "401":
          description: Missing or incorrect `X-API-Key` header.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Unauthorized"
        "404":
          description: >
            No JustWatch page could be found or parsed for `title` (e.g.
            the guessed slug 404s, or the title is misspelled).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FilmResult"
              example:
                title: "Not A Real Movie"
                service: []
                error: "No JustWatch page found at https://www.justwatch.com/us/movie/not-a-real-movie"
        "500":
          description: >
            The server has no `ADMIN_API_KEY` configured, so no key could
            possibly be valid. A deployment/configuration problem, not a
            caller error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "ADMIN_API_KEY is not configured on the server"
        "502":
          description: >
            The request to JustWatch itself failed (timeout, connection
            error, DNS failure, non-404 HTTP error, etc.) -- distinct from
            404, which means JustWatch responded but the title genuinely
            wasn't found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FilmResult"
              example:
                title: "Blood Feast"
                service: []
                error: "HTTPSConnectionPool(host='www.justwatch.com', port=443): Read timed out."
    post:
      operationId: scrapeBatch
      tags: [Scrape]
      summary: Scrape a caller-supplied list of films
      description: >
        Stateless: this service stores no film list of its own, so the
        caller supplies one on every request. Returns one `FilmResult`
        per input film, in the same order, via a bounded-concurrency
        path (`scrape_titles()`) -- one invocation handles the whole
        list rather than one round trip per film. A single film failing
        (404, bad `justwatch_url`, network error) doesn't fail the
        request; that film's result just carries `service: []` and an
        `error`.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              minItems: 1
              items:
                $ref: "#/components/schemas/ScrapeBatchFilm"
            example:
              - title: "The Ring"
                justwatch_url: "https://www.justwatch.com/us/movie/le-cercle"
              - title: "Hereditary"
      responses:
        "200":
          description: One FilmResult per input film, in the same order as the request.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FilmResult"
        "400":
          description: >
            Malformed request: body isn't a JSON array, the array is
            empty, an item isn't a JSON object, or an item is missing
            `title`. A bad `justwatch_url` is *not* a 400 here -- it
            fails only that one film (see the `200` response), the same
            way `?url=` failing fails only that one request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or incorrect `X-API-Key` header.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: The server has no `ADMIN_API_KEY` configured.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /films:
    get:
      operationId: listFilms
      tags: [Films]
      summary: List the film catalogue, each film with its stored offers
      description: >
        Returns every film with its nested `services` (offers), ordered by
        year then the catalogue's own sort order. This is the Supabase-backed
        stored state, not a live scrape -- offers are whatever the last
        refresh persisted.


        Requires `X-API-Key`, like the writes -- but at the read tier, so
        `READ_API_KEY` is accepted here as well as `ADMIN_API_KEY`. The
        intended caller is a server-side consumer holding a key; there are no
        CORS headers here, so a browser can't call this directly in any case.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: year
          in: query
          required: false
          description: Restrict to a single release year.
          schema:
            type: integer
          example: 2024
      responses:
        "200":
          description: The catalogue (optionally filtered by year).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Film"
        "400":
          description: "`year` was provided but isn't an integer."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or incorrect `X-API-Key` header.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Supabase isn't configured on the server (missing URL / service-role key).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Supabase couldn't be reached or returned an error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createFilm
      tags: [Films]
      summary: Add a film to the catalogue
      description: >
        Admin only. Creates a film row; offers are populated separately via
        refresh. `year`+`title` must be unique -- a duplicate returns 409.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FilmCreate"
      responses:
        "201":
          description: The created film.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Film"
        "400":
          description: >
            Missing or invalid fields (year, date, title are required;
            justwatch_url, if given, must be a JustWatch URL).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or incorrect `X-API-Key` header.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: A film with this (year, title) already exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: The server has no `ADMIN_API_KEY`, or Supabase isn't configured.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Supabase couldn't be reached or returned an error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /films/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
        example: 20
    get:
      operationId: getFilm
      tags: [Films]
      summary: Get one film with its offers
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: The film with its nested offers.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Film"
        "401":
          description: Missing or incorrect `X-API-Key` header.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No film with that id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Supabase isn't configured on the server.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Supabase couldn't be reached or returned an error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateFilm
      tags: [Films]
      summary: Update a film's fields (including the justwatch_url override)
      description: >
        Admin only. Send any subset of the writable fields. Set `justwatch_url`
        to a JustWatch URL to override slug-guessing (the "Override"), or to
        null to clear it. The change takes effect on the next refresh.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FilmPatch"
            example:
              justwatch_url: "https://www.justwatch.com/us/movie/le-cercle"
      responses:
        "200":
          description: The updated film.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Film"
        "400":
          description: Empty body, no updatable fields, or an invalid field value.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or incorrect `X-API-Key` header.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No film with that id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: The server has no `ADMIN_API_KEY`, or Supabase isn't configured.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Supabase couldn't be reached or returned an error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteFilm
      tags: [Films]
      summary: Delete a film (its offers cascade)
      security:
        - ApiKeyAuth: []
      responses:
        "204":
          description: Deleted.
        "401":
          description: Missing or incorrect `X-API-Key` header.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No film with that id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: The server has no `ADMIN_API_KEY`, or Supabase isn't configured.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Supabase couldn't be reached or returned an error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /films/{id}/refresh:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
        example: 20
    post:
      operationId: refreshFilm
      tags: [Films]
      summary: Re-scrape one film and replace its stored offers
      description: >
        Admin only. Scrapes the film's JustWatch page -- using its stored
        `justwatch_url` override if set, else the slug guessed from the title
        -- and replaces that film's offers in Supabase. On a scrape/network
        failure the existing offers are left untouched (a transient upstream
        failure never wipes good data) and the endpoint returns 502.
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Offers refreshed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshResult"
              example:
                film_id: 20
                title: "The Ring"
                url: "https://www.justwatch.com/us/movie/le-cercle"
                services_updated: 16
                refreshed: true
        "401":
          description: Missing or incorrect `X-API-Key` header.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No film with that id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: The server has no `ADMIN_API_KEY`, or Supabase isn't configured.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: >
            The scrape failed (JustWatch timeout / 404 / parse error) or
            Supabase errored. Existing offers are left unchanged.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshResult"
              example:
                film_id: 20
                title: "The Ring"
                error: "No JustWatch page found at https://www.justwatch.com/us/movie/the-ring"
                refreshed: false
                note: "existing offers left unchanged"
  /films/refresh:
    post:
      operationId: refreshAllFilms
      tags: [Films]
      summary: Re-scrape every film and refresh its offers
      description: >
        Admin only. The periodic job (e.g. a scheduled Vercel Cron): scrapes
        all films with bounded concurrency and replaces each film's offers.
        One film's scrape failure doesn't abort the batch or wipe that film's
        offers -- it's reported inline with `refreshed: false`.


        Guarded against a broken scrape: if every film that could be read came
        back with zero offers, the whole run is refused with `502` and nothing
        is written. A handful of films genuinely stream nowhere, but all of
        them doing so at once means the scrape failed, not that the catalogue
        emptied overnight. Films that failed to scrape outright don't count
        toward this -- they were never read, so there's nothing to refuse.
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Per-film refresh summary.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshBatchResult"
        "401":
          description: Missing or incorrect `X-API-Key` header.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: The server has no `ADMIN_API_KEY`, or Supabase isn't configured.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: >
            Supabase couldn't be reached, listing the catalogue failed, or the
            run was refused because every readable film scraped empty. In the
            refusal case nothing was written and `refreshed` is `0`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: >
        Must match the server's `ADMIN_API_KEY` environment variable, which
        authorises everything including writes and refreshes.


        The two read endpoints (`GET /films`, `GET /films/{id}`) additionally
        accept a read-only key, `READ_API_KEY`, so catalogue access can be
        granted without granting the ability to change anything. That key is
        optional and unset by default: with none configured, reads accept the
        admin key alone. Paste whichever key you hold into Authorize — the
        header is the same either way.


        Note that a browser-based third-party client would also need CORS
        headers, which this service deliberately does not send.
  schemas:
...
    FilmResult:
      type: object
      description: >
        Result of scraping a single film (the `?title=` response shape).
        `url` is only present when the scrape succeeded (200) -- it's
        omitted from 404/502 error bodies, which only have `title`,
        `service`, and `error`.
      required: [title, service]
      properties:
        title:
          type: string
        url:
          type: string
          format: uri
          description: The canonical JustWatch URL that was actually scraped. Only present on success.
        service:
          type: array
          items:
            $ref: "#/components/schemas/Offer"
        error:
          type: string
          nullable: true
          description: Present only on failure; `service` will be empty.
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...